Fixes regarding the wikibase-sync dependencies
Due to the impossibility of updating the PyPi package of wikibase-sync at the moment, the library has been manuallly added in a sub-directory.
The installation has to be invoked manually, but all the dependencies with the necessary versions are already present in the requirements.txt file.
The .bat file has been updated and simplified.

When the script is run, it is able to crete most of the environment, but one of the python files (wikibase_config.py) raises an exception as it is not able to connect to the sparql endpoint.